Fix sample
This commit tries to clarify the sample with regards to some discussion
on the Spring Boot issue tracker, in particular:

https://github.com/spring-projects/spring-boot/issues/8275

To be able to reuse an `HazelcastInstance` in server mode, the location
of the configuration file should be set in both auto-configuration
unfortunately: there is currently now way to reuse an instance that was
loaded via the default configuration. See the following issue:

https://github.com/hazelcast/hazelcast/issues/10007

In a client scenario, the only way is to effectively disable the
auto-configuration which annoys me a lot. Probably Spring Boot should
back off automatically in such a scenario but we need a way to reliably
figure this out.

Also, I wasn't aware that Hazelcast provides different `CachingProvider`
that are looking for different things. IMO, there is a gap in the
caching support in Spring Boot about it and I'd be more than happy to
improve it as long as I don't need to resort to Hazelcast specific API
inside the JCache auto-configuration