Do not use deprecated methods of ObjectDataInput/Output, PortableReader/Writer, and ClassDefinitionBuilder (#470)
We have deprecated methods containing `UTF` in their names and
introduced new methods that contain `String` instead of it in https://github.com/hazelcast/hazelcast/pull/18100.

This is a cleanup PR to use newly introduced methods instead of the
deprecated ones. The change is done automatically via IDE and
verified manually by going over the changes.

Also, fixed compilation errors due to the changes in the GenericRecords.