Make default Hazelcast version, last released version (i.e. not `SNAPSHOT`).
[`SNAPSHOT` Hazelcast builds are no longer publically
accessible](https://hazelcast.atlassian.net/browse/DI-100), which leads
to build failures as `code-samples` relies on these versions.

It makes more sense for examples of how to use Hazelcast to be backed by
production-quality dependencies, i.e. a released version.

As such, downgraded to the latest released version of Hazelcast - will
be kept up-to-date via
[Dependabot](https://github.com/hazelcast/hazelcast-code-samples/pull/574).

[Slack discussion](https://hazelcast.slack.com/archives/C05LM8B80UT/p1716380405098929)