Move Spring examples to a separate directory [DEX-294] (#688)
The `hazelcast-integration` directory contains examples that were in the
scope of the Integration team.

From users' POV it doesn't make sense to structure the examples in a way
that reflects our internal structure. We should rather structure them by
topics.

This only moves the examples and updates pom structure. I will fix and
update the examples in separate PRs.